added screen02, edit source code Main.java

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int balance = 2000000000;
-        int reffil = 500000000;
-        int total = balance + reffil;
+        long balance = 2000000000;
+        long reffil = 500000000;
+        long total = (balance + reffil);
         System.out.println(total);
     }
 }
